python-services/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Rate limiting
slowapi>=0.1.9

# AWS (SQS worker + S3 for file storage)
boto3>=1.34.0

# Lambda adapter
mangum>=0.17.0

# Document Processing
pypdf2>=3.0.0
python-docx>=1.0.0
langchain>=0.1.0
langchain-text-splitters>=0.1.0

# Vector Database & Embeddings
pinecone>=2.2.0
openai>=1.0.0
tiktoken>=0.5.0
sentence-transformers>=2.2.0

# Database
asyncpg>=0.28.0
sqlalchemy>=2.0.0

# Caching & Background Tasks
redis>=5.0.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.24.0
numpy>=1.24.0
pandas>=2.0.0
